Semantic_Segmentation/requirements.txt
tqdm
numpy
visdom
pillow
torchvision